# make putNode correct 