--- a/skl/arena.go
+++ b/skl/arena.go
@@ -35,9 +35,8 @@ const (
 
 // Arena should be lock-free.
 type Arena struct {
-	n             uint32
-	buf           []byte
-	buf0Alignment uint32
+	n   uint32
+	buf []byte
 }
 
 // newArena returns a new arena.
@@ -46,10 +45,16 @@ func newArena(n int64) *Arena {
 	// of nil pointer.
 	out := &Arena{
 		n:   1,
-		buf: make([]byte, n),
+		buf: makeAlignedBuf(n),
 	}
-	out.buf0Alignment = uint32(uintptr(unsafe.Pointer(&out.buf[0]))) & ^uint32(nodeAlign)
 	return out
+}
+
+func makeAlignedBuf(n int64) []byte {
+	buf := make([]byte, n+int64(nodeAlign))
+	buf0Alignment := uint32(uintptr(unsafe.Pointer(&buf[0]))) & ^uint32(nodeAlign)
+	buf = buf[buf0Alignment : int64(buf0Alignment)+n]
+	return buf
 }
 
 func (s *Arena) size() int64 {
@@ -75,8 +80,7 @@ func (s *Arena) putNode(height int) uint32 {
 		l, n, len(s.buf))
 
 	// Return the aligned offset.
-	m := (s.buf0Alignment + (n - l + uint32(nodeAlign))) & ^uint32(nodeAlign)
-	m -= s.buf0Alignment
+	m := (n - l + uint32(nodeAlign)) & ^uint32(nodeAlign)
 	return m
 }