`s.buf` bytes may be not aligned, so the correct implementation should take that into consideration.

**Update** 

Switched to use `makeAlignedBuf` instead.
<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/dgraph-io/badger/1272)
<!-- Reviewable:end -->