@zhiqiangxu 2020-03-24
node address = address of first element + offset

In order to make sure node address is aligned, should consider `address of first element`, which is not necessarily aligned.