@zhiqiangxu 2020-03-25
I missed one thing shamefully: 

The first word in a variable or in an allocated struct, array, or slice can be relied upon to be 64-bit aligned.