@zhiqiangxu 2020-03-25
The spec only guarantees :

the first (64-bit) word in a variable or in an allocated struct, array, or slice can be relied upon to be 64-bit aligned.

So it's not guaranteed that the first element of a byte slice is 8 byte aligned.