@jarifibrahim 2020-05-26
@zhiqiangxu I did a small benchmark to compare the performance.
This is the `BenchmarkWrite` function in the `skl` package.
```
name            old time/op    new time/op    delta
Write-8          522ns ± 8%     563ns ± 3%  +7.87%  (p=0.001 n=10+8)
```

The new changes seem to be slower. Do you know why it could be slower?