@zhiqiangxu 2020-06-01
Maybe because of the extra [bounding check](https://go101.org/article/bounds-check-elimination.html) involved, but pity that I don't know how to optimize it:(