@damz 2020-07-02
The rationale looks correct to me: Golang doesn't guarantee that the underlying memory block of a `[]byte` is aligned. It only guarantees that the underlying memory block of an array follows the alignment guarantees of each of the element:

From https://golang.org/ref/spec#Size_and_alignment_guarantees:

> For a variable x of array type: unsafe.Alignof(x) is the same as the alignment of a variable of the array's element type.

So a `[]uint64` will be aligned to 64 bit on platforms that requires that. But a `[]byte` will not.

Based on this, I can suggest the following:

```golang
func makeAligned(n int64) []byte {
	buf := make([]uint64, n>>3)
	head := (*reflect.SliceHeader)(unsafe.Pointer(&buf))
	head.Len = int(n)
	head.Cap = int(n)
	return *(*[]byte)(unsafe.Pointer(head))
}
```

It doesn't seem to have any measurable performance impact.