@harshil-goel 2024-10-14
From what I can understand we already do bit alignment while we are inserting data using putNode or putVal. They ensure alignment before inserting. Hence alignment while creating the buffer is not required. Let me know if you belive so and reopen the PR.